requirements.txt
django
pillow
django-summernote
django-smart-selects
